backend/requirements.txt
Flask>=2.2
Flask-Migrate>=4.0
Flask-SQLAlchemy>=3.0
PyMySQL>=1.0
python-dotenv>=1.0
pytest>=7.0
black
flake8
bcrypt
